+++ lib/src/internal/page_widget.dart
@@ -4,6 +4,7 @@ import 'package:demoflu/src/page.dart';
 import 'package:demoflu/src/internal/model.dart';
 import 'package:demoflu/src/internal/provider.dart';
 import 'package:demoflu/src/sections/page_section.dart';
+import 'package:demoflu/src/sections/space_section.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
@@ -18,21 +19,24 @@ class DemoFluPageWidget extends StatelessWidget {
     List<Widget> children = [];
 
     final DemoMenuItem menuItem = model.selectedMenuItem;
+    SpaceSection spaceSection = SpaceSection();
     if (menuItem.page != null) {
       final DemoFluPage page = menuItem.page!();
 
       children.add(BreadcrumbWidget(menuItem: menuItem));
       for (PageSection section in PageHelper.sectionsFrom(page)) {
+        if (page.autoSpace) {
+          children.add(spaceSection.buildWidget(context));
+        }
         children.add(section.buildWidget(context));
       }
     }
 
-    return LayoutBuilder(builder: (context, constraints) {
-      return ListView.separated(
-          padding: EdgeInsets.fromLTRB(32, 16, 32, 32),
-          itemBuilder: (context, index) => children[index],
-          separatorBuilder: (context, index) => SizedBox(height: 24),
-          itemCount: children.length);
-    });
+    return SingleChildScrollView(
+        child: Padding(
+            padding: EdgeInsets.fromLTRB(32, 16, 32, 32),
+            child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
+                children: children)));
   }
 }