+++ lib/src/internal/sections/widget_section_ui.dart
@@ -1,12 +1,12 @@
-import 'package:demoflu/src/internal/sections/titled_widget.dart';
+import 'package:demoflu/src/internal/sections/titled_section_ui.dart';
 import 'package:demoflu/src/sections/widget_section.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 
 /// Container for widget example session
 @internal
-class WidgetContainer extends StatelessWidget {
-  WidgetContainer(this.section);
+class WidgetSectionUI extends StatelessWidget {
+  WidgetSectionUI(this.section);
 
   final WidgetSection section;
 
@@ -25,7 +25,7 @@ class WidgetContainer extends StatelessWidget {
   Widget _container(BuildContext context) {
     return Align(
         alignment: Alignment.centerLeft,
-        child: TitledWidget(
+        child: TitledSectionUI(
             child: section.widgetBuilder(context), section: section));
   }
 }