+++ lib/src/sections/space_section.dart
@@ -1,14 +1,18 @@
+import 'package:demoflu/src/internal/provider.dart';
 import 'package:demoflu/src/sections/page_section.dart';
+import 'package:demoflu/src/sections_defaults.dart';
 import 'package:flutter/widgets.dart';
 
 class SpaceSection extends PageSection {
-  SpaceSection({required double height}) : _height = height >= 0 ? height : 0;
+  SpaceSection({this.height});
 
-  double _height;
-  double get height => _height;
+  double? height;
 
   @override
   Widget buildWidget(BuildContext context) {
+    SectionDefaults defaults = DemoFluProvider.sectionDefaultsOf(context);
+    double height = this.height ?? defaults.spaceHeight;
+    height = height >= 0 ? height : 0;
     return SizedBox(height: height);
   }
 }