+++ lib/src/sections/widget_section.dart
@@ -1,4 +1,4 @@
-import 'package:demoflu/src/internal/sections/example_widget.dart';
+import 'package:demoflu/src/internal/sections/widget_section_ui.dart';
 import 'package:demoflu/src/sections/titled_page_section.dart';
 import 'package:flutter/widgets.dart';
 
@@ -21,6 +21,6 @@ class WidgetSection extends TitledPageSection {
 
   @override
   Widget buildWidget(BuildContext context) {
-    return WidgetContainer(this);
+    return WidgetSectionUI(this);
   }
 }